requirements.txt
asgiref==3.8.1
Django==5.2.1
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.19.16
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
whitenoise==6.9.0
